% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Absence_Check.R
\name{Absence_Check}
\alias{Absence_Check}
\title{Absence Checker}
\usage{
Absence_Check(
  df,
  assumed_zero_table = Absence_Table,
  assumed_zero_absence_inputs = c("a"),
  method = "check",
  CPC_Code_Column = "CPC_Code",
  comments = TRUE,
  comment_col = "comments",
  Water_check = TRUE,
  water_column = "WATERg",
  CHOAVL_check = TRUE,
  CHOAVL_column = "CHOAVLg",
  CHOAVLDF_check = TRUE,
  CHOAVLDFg_std_column = "CHOAVLDFg_standardised",
  ALC_check = TRUE,
  ALC_column = "ALCg",
  FIBG_check = TRUE,
  FIBG_column = "FIBGg",
  NIA_check = TRUE,
  NIA_column = "NIAmg",
  VITB6_check = TRUE,
  VITB6_column = "VITB6_mg_standardised",
  VITB12_check = TRUE,
  VITB12_column = "VITB12mcg",
  CU_check = TRUE,
  CU_column = "CUmg",
  F22D6N3_check = TRUE,
  F22D6N3_column = "F22D6N3g",
  F20D5N3_check = TRUE,
  F20D5N3_column = "F20D5N3g",
  assumed_zero_water_col = "WATERg",
  assumed_zero_CHOAVL_col = "CHOAVLg",
  assumed_zero_CHOAVLDF_col = "CHOAVLDFg_standardised",
  assumed_zero_ALC_col = "ALCg",
  assumed_zero_FIBG_col = "FIBGg",
  assumed_zero_NIA_col = "NIAmg",
  assumed_zero_VITB6_col = "VITB6_mg_standardised",
  assumed_zero_VITB12_col = "VITB12mg",
  assumed_zero_F22D6N3_col = "F22D6N3g",
  assumed_zero_F20D5N3_col = "F20D5N3g",
  assumed_zero_CU_col = "CUmg",
  assumed_zero_CPC_column = "CPC_Code",
  No_Water_CPC_Codes = c(""),
  No_CHOAVL_CPC_Codes = c(""),
  No_CHOAVLDFg_standardised_CPC_Codes = c(""),
  No_ALC_CPC_Codes = c(""),
  No_FIBG_CPC_Codes = c(""),
  No_NIA_CPC_Codes = c(""),
  No_VITB6_CPC_Codes = c(""),
  No_VITB12_CPC_Codes = c(""),
  No_F22D6N3_CPC_Codes = c(""),
  No_F20D5N3_CPC_Codes = c(""),
  No_CU_CPC_Codes = c("")
)
}
\arguments{
\item{df}{Required - The data.frame which contains the food items you are
looking to check.}

\item{assumed_zero_table}{Optional - default: \code{Absence_Table} - Choose
whether to use the supplied preset table which contains information on
which CPC code shouldn't contain certain nutrients.}

\item{assumed_zero_absence_inputs}{Optional - default: \code{c("a")}
\itemize{
\item If using the default \code{assumed_zero_table}, this allows you to pick
which options for selection you want. Choosing \code{c("a")} only applies
corrections that are certain, \code{c("a", "b")} results in corrections
being applied to a broader selection of checks and food groups, with high
certainty, but less certain than just using \code{c("a")}.
}}

\item{method}{Required - default: \code{"check"} - Either \code{"check"},
\code{"fill_all"},  or \code{"fill_blank"}. If set to \code{"check"}, the
function will print out warnings where it encounters values which should be
zero. If set to \code{"fill_all"}, the function will correct any incorrect
values to 0. If set to \code{"fill_blank"}, the function will correct
incorrect values to 0 only if they're missing (e.g. NA or "").}

\item{CPC_Code_Column}{Required - default: \code{"CPC_Code"} - the name of
the column in \code{df} which contains the CPC codes for the food items.}

\item{comments}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Choose whether to record the changes made. It is strongly
recommended to leave this as \code{TRUE} to maintain records of any changes
in the data.}

\item{comment_col}{Optional - default: \code{"comments"} - Required if
\code{comments} is set to \code{TRUE}. The name of the column in \code{df}
where comments should be stored. If the entry for \code{comment_col} is not
the name of a column in \code{df}, then a new column will be created with
the name of the entry.}

\item{Water_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Water values in
\code{df}.}

\item{water_column}{Optional - default: \code{"WATERg"} - Required if
\code{comments} is set to \code{TRUE}. The name of the column in \code{df}
which contains water values.}

\item{CHOAVL_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Available
Carbohydrate values in \code{df}.}

\item{CHOAVL_column}{Optional - default: \code{"CHOAVLg"} - Required if
\code{CHOAVL_check} is set to \code{TRUE}. The name of the column in
\code{df} which contains Available Carbohydrate values.}

\item{CHOAVLDF_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Available
Carbohydrate (calculated by difference) values in \code{df}.}

\item{CHOAVLDFg_std_column}{Optional - default:
\code{"CHOAVLDFg_standardised"} - Required if \code{CHOAVLDF_check} is set
to \code{TRUE}. The name of the column in \code{df} which contains
Available Carbohydrate (calculated by difference) values.}

\item{ALC_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Alcohol values in
\code{df}.}

\item{ALC_column}{Optional - default: \code{"ALCg"} - Required if
\code{ALC_check} is set to \code{TRUE}. The name of the column in \code{df}
which contains Alcohol values.}

\item{FIBG_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Fibre values in
\code{df}.}

\item{FIBG_column}{Optional - default: \code{"FIBGg"} - Required if
\code{FIBG_check} is set to \code{TRUE}. The name of the column in
\code{df} which contains Fibre values.}

\item{NIA_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Niacin values in
\code{df}.}

\item{NIA_column}{Optional - default: \code{"NIAmg"} - Required if
\code{NIA_check} is set to \code{TRUE}. The name of the column in \code{df}
which contains Niacin values.}

\item{VITB6_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Vitamin B6 values
in \code{df}.}

\item{VITB6_column}{Optional - default: \code{"VITB6_mg_standardised"} -
Required if \code{VITB6_check} is set to \code{TRUE}. The name of the
column in \code{df} which contains Vitamin B6 values.}

\item{VITB12_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Vitamin B12 values
in \code{df}.}

\item{VITB12_column}{Optional - default: \code{"VITB12mcg"} - Required if
\code{VITB12_check} is set to \code{TRUE}. The name of the column in
\code{df} which contains Vitamin B12 values.}

\item{CU_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the Copper values in
\code{df}.}

\item{CU_column}{Optional - default: \code{"CUmg"} - Required if
\code{CU_check} is set to \code{TRUE}. The name of the column in \code{df}
which contains Copper values.}

\item{F22D6N3_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the F22D6N3 values in
\code{df}.}

\item{F22D6N3_column}{Optional - default: \code{"F22D6N3g"} - Required if
\code{F22D6N3_check} is set to \code{TRUE}. The name of the column in
\code{df} which contains F22D6N3 values.}

\item{F20D5N3_check}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. Decide whether you would like to check the F20D5N3 values in
\code{df}.}

\item{F20D5N3_column}{Optional - default: \code{"F20D5N3g"} - Required if
\code{F20D5N3_check} is set to \code{TRUE}. The name of the column in
\code{df} which contains F20D5N3 values.}

\item{assumed_zero_water_col}{Optional - default: \code{"WATERg"} - The
column in the \code{assumed_zero_table} that contains the absence status of
water for different food groups. If using \code{assumed_zero_table}, and if
using the default table, there is no need to change this. If using a
modified table, please change this to match the column that lists water in
that modified table.}

\item{assumed_zero_CHOAVL_col}{Optional - default: \code{"CHOAVLg"} - The
column in the \code{assumed_zero_table} that contains the absence status of
Carbohydrates for different food groups. If using
\code{assumed_zero_table}, and if using the default table, there is no need
to change this. If using a modified table, please change this to match the
column that lists Carbohydrates in that modified table.}

\item{assumed_zero_CHOAVLDF_col}{Optional - default:
\code{"CHOAVLDFg_standardised"} - The column in the
\code{assumed_zero_table} that contains the absence status of Carbohydrates
calculated by difference for different food groups. If using
\code{assumed_zero_table}, and if using the default table, there is no need
to change this. If using a modified table, please change this to match the
column that lists Carbohydrates calculated by difference in that modified
table.}

\item{assumed_zero_ALC_col}{Optional - default: \code{"ALCg"} - The column in
the \code{assumed_zero_table} that contains the absence status of alcohol
for different food groups. If using \code{assumed_zero_table}, and if using
the default table, there is no need to change this. If using a modified
table, please change this to match the column that lists Alcohol in that
modified table.}

\item{assumed_zero_FIBG_col}{Optional - default: \code{"FIBGg"} - The column
in the \code{assumed_zero_table} that contains the absence status of Fibre
for different food groups. If using \code{assumed_zero_table}, and if using
the default table, there is no need to change this. If using a modified
table, please change this to match the column that lists Fibre in that
modified table.}

\item{assumed_zero_NIA_col}{Optional - default: \code{"NIAmg"} - The column
in the \code{assumed_zero_table} that contains the absence status of Niacin
for different food groups. If using \code{assumed_zero_table}, and if using
the default table, there is no need to change this. If using a modified
table, please change this to match the column that lists Niacin in that
modified table.}

\item{assumed_zero_VITB6_col}{Optional - default:
\code{"VITB6_mg_standardised"} - The column in the
\code{assumed_zero_table} that contains the absence status of Vitamin B6
for different food groups. If using \code{assumed_zero_table}, and if using
the default table, there is no need to change this. If using a modified
table, please change this to match the column that lists Vitamin B6 in that
modified table.}

\item{assumed_zero_VITB12_col}{Optional - default: \code{"VITB12mg"} - The
column in the \code{assumed_zero_table} that contains the absence status of
Vitamin B12 for different food groups. If using \code{assumed_zero_table},
and if using the default table, there is no need to change this. If using a
modified table, please change this to match the column that lists Vitamin
B12 in that modified table.}

\item{assumed_zero_F22D6N3_col}{Optional - default: \code{"F22D6N3g"} - The
column in the \code{assumed_zero_table} that contains the absence status of
F22D6N3 for different food groups. If using \code{assumed_zero_table}, and
if using the default table, there is no need to change this. If using a
modified table, please change this to match the column that lists F22D6N3
in that modified table.}

\item{assumed_zero_F20D5N3_col}{Optional - default: \code{"F20D5N3g"} - The
column in the \code{assumed_zero_table} that contains the absence status of
F20D5N3 for different food groups. If using \code{assumed_zero_table}, and
if using the default table, there is no need to change this. If using a
modified table, please change this to match the column that lists F20D5N3
in that modified table.}

\item{assumed_zero_CU_col}{Optional - default: \code{"CUmg"} - The column in
the \code{assumed_zero_table} that contains the absence status of Copper
for different food groups. If using \code{assumed_zero_table}, and if using
the default table, there is no need to change this. If using a modified
table, please change this to match the column that lists Copper in that
modified table.}

\item{assumed_zero_CPC_column}{Optional - default: \code{"CPC_Code"} - The
column in the \code{assumed_zero_table} that contains the CPC Codes for
different food groups. If using \code{assumed_zero_table}, and if using the
default table, there is no need to change this. If using a modified table,
please change this to match the column that lists the CPC Codes in that
modified table.}

\item{No_Water_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{water_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Water.}

\item{No_CHOAVL_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{CHOAVL_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Carbohydrates.}

\item{No_CHOAVLDFg_standardised_CPC_Codes}{Optional - default: \code{c("")} -
Required if not using the \code{assumed_zero_table} input, if
\code{CHOAVLDF_check} is \code{TRUE}. Please input the CPC codes where you
wouldn't expect to find Carbohydrates calculated by difference.}

\item{No_ALC_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{ALC_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Alcohol.}

\item{No_FIBG_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{FIBG_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Fibre.}

\item{No_NIA_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{NIA_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Niacin.}

\item{No_VITB6_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{VITB6_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Vitamin B6.}

\item{No_VITB12_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{VITB12_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Vitamin B12.}

\item{No_F22D6N3_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{F22D6N3_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
F22D6N3.}

\item{No_F20D5N3_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{F20D5N3_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
F20D5N3.}

\item{No_CU_CPC_Codes}{Optional - default: \code{c("")} - Required if not
using the \code{assumed_zero_table} input, if \code{CU_check} is
\code{TRUE}. Please input the CPC codes where you wouldn't expect to find
Copper.}
}
\value{
If \code{method = "check"}, the function will output a list of
problematic entries in the data.frame to the console for manual
investigation. If \code{method = "fill_all"}, the output will be a
data.frame that mimics the input \code{df}, but with nutrient values
changed where an error is detected - e.g. if a food item is in a CPC group
that shouldn't contain water, and it does, then the water value for that
item will be changed to 0. If \code{method = "fill_blank"}, the output will
be a data.frame that mimics the input \code{df}, but with nutrient values
changed where blank value is detect where we can assume the value to be 0 -
e.g. if a food item is in a CPC group that shouldn't contain water, and the
water value for that item is NA, or "", then the water value for that item
will be changed to 0.
}
\description{
The Absence Checker runs over a data.frame containing food items
with CPC codes, and checks those food items and the nutrients they contain
against a list. If a nutrient is showing as present when that food item
shouldn't contain that nutrient, the Absence_Check function will sort those
values, depending on the user input.
}
