% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breakfast_df}
\alias{breakfast_df}
\title{Imaginary Food Items}
\format{
A data frame with 10 rows and 21 columns:
\describe{
\item{food_code}{Imaginary Food ID}
\item{food_name}{Food names to attach the imaginary nutritional data to}
\item{WATERg}{Imaginary Water values}
\item{PROCNTg}{Imaginary Protein values}
\item{FAT_g_combined}{Imaginary Fat values}
\item{CHOAVLg}{Imaginary Carbohyrate values}
\item{FIBTGg_combined}{Imaginary Fibre values}
\item{ALCg}{Imaginary Alcohol values}
\item{ASHg}{Imaginary Ash values}
\item{THIAmg}{Imaginary Thiamine values}
\item{THIAHCLmg}{Imaginary Thiamine from HCL values}
\item{RETOLmcg}{Imaginary retinol values}
\item{CARTBEQmcg_combined}{Imaginary Beta-Carotene equiv. values}
\item{NIAmg}{Imaginary Niacin values}
\item{TRPmg}{Imaginary Tryptophan values}
\item{NIAEQmg}{Imaginary Niacin equivalent values}
\item{NIATRPmg}{Imaginary Niacin from Tryptophan values}
\item{FATg}{Imaginary Fat values}
\item{FAT_g}{Imaginary Fat from unknown source values}
\item{FATCEg}{Imaginary Fat from CE method values}
\item{comments}{Imaginary comments}
}
}
\source{
An imaginary dataset; no formal source.
}
\usage{
breakfast_df
}
\description{
An imaginary dataset of Food Items used to demonstrate various functions
}
\keyword{datasets}
