% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breakfast_df_nonstandard}
\alias{breakfast_df_nonstandard}
\title{Imaginary Food Items}
\format{
A data frame with 10 rows and 21 columns:
\describe{
  \item{food_code}{Imaginary Food ID}
  \item{food_name}{Food names to attach the imaginary nutritional data to}
  \item{Water_values_g}{Imaginary Water values}
  \item{PROCNT_values_g}{Imaginary Protein values}
  \item{FAT_values_g_combined}{Imaginary Fat values}
  \item{CHOAVL_values_g}{Imaginary Carbohyrate values}
  \item{FIBTG_values_g_combined}{Imaginary Fibre values}
  \item{ALC_values_g}{Imaginary Alcohol values}
  \item{ASH_values_g}{Imaginary Ash values}
  \item{Thiamine_milligrams}{Imaginary Thiamine values}
  \item{Thiamine_from_HCL_milligrams}{Imaginary Thiamine from HCL values}
  \item{Retinol_micrograms}{Imaginary retinol values}
  \item{Beta_Carotene_Equivalents_micrograms}{Imaginary Beta-Carotene equiv. values}
  \item{Niacin_milligrams}{Imaginary Niacin values}
  \item{Tryptophan_milligrams}{Imaginary Tryptophan values}
  \item{Niacin_eq_milligrams}{Imaginary Niacin equivalent values}
  \item{Niacine_from_TRP_mg}{Imaginary Niacin from Tryptophan values}
  \item{FAT_in_g}{Imaginary Fat values}
  \item{FAT_unknown_calc_g}{Imaginary Fat from unknown source values}
  \item{FAT_continuous_extraction_g}{Imaginary Fat from CE method values}
  \item{comments_column}{Imaginary comments}
}
}
\source{
An imaginary dataset; no formal source.
}
\usage{
breakfast_df_nonstandard
}
\description{
An imaginary dataset of Food Items with non-standard column names used to
demonstrate various functions
}
\keyword{datasets}
