% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change_Summary.R
\name{Change_Summary}
\alias{Change_Summary}
\title{Change Summary}
\usage{
Change_Summary(df, comment_col = "comments", detailed = FALSE)
}
\arguments{
\item{df}{Required - the data.frame which contains the food items that have
been changed by functions from NutritionTools.}

\item{comment_col}{Required - default: \code{'comments'} - The name of the
column within \code{df} that contains comments.}

\item{detailed}{Required - default: \code{FALSE} - \code{TRUE} or
\code{FALSE}. If set to \code{TRUE}, then the Change Summary will examine
the exact comments as they appear in the data frame. If set to
\code{FALSE}, as it is by default, comments which have unique identifiers
(values, imputation IDs etc) have those aspects removed, to reduce the
number of return types.}
}
\value{
A console output detailing how many times one of the comments
generated by functions in NutritionTools appears in the df.
}
\description{
This function is designed to be run on a data frame after other functions
from the NutritionTools package have been run on it. This function scans the
comments column for the preset comments left by other functions from the
package, and produces a summary of them, and how often they appear.
}
