% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Column_Comparison.R
\name{Column_Comparison}
\alias{Column_Comparison}
\title{Column Name Comparer}
\usage{
Column_Comparison(df1, df2, Summary_Message = TRUE)
}
\arguments{
\item{df1}{Required - the first data.frame to be compared.}

\item{df2}{Required - the second data.frame to be compared.}

\item{Summary_Message}{Optional - default: \code{TRUE} - \code{TRUE} or
\code{FALSE}. Decides whether a summary message should be included.}
}
\value{
A comparison data.frame, listing the names of the columns from df1
  and df2.
}
\description{
Compares the column names of the two input data.frames,
  \code{df1} and \code{df2}, and displays whether they are the same or not.
  Useful ahead of an rbind function. If the columns are uneven, the function
  will create dummy columns to make up the numbers for the comparison.
}
\examples{
# Example data.frames have been created to give an example of using the
# Column_Comparison to compare the column names.

breakfast_df
#
breakfast_df_nonstandard

# We will start with two data.frames who's names do not line up.

Col_Comparison <- Column_Comparison(breakfast_df, breakfast_df_nonstandard)

Col_Comparison

# Note how most of the columns do not match in their names, and the function
# tells you so.


# If the data.frames don't line up in terms of column numbers, then filler
# columns will be added.

breakfast_df_2 <- breakfast_df
breakfast_df_2$THIAmg <- NULL
breakfast_df_2$TRPmg <- NULL
breakfast_df_2$FAT_g <- NULL

Col_Comparison <- Column_Comparison(breakfast_df, breakfast_df_2)

Col_Comparison

# Note how new columns are added to make up the numbers - however, these
# columns are added to the end of the data.frame, causing a shift.

# If the columns line up perfectly however:

breakfast_df_copy <- breakfast_df

Col_Comparison <- Column_Comparison(breakfast_df, breakfast_df_copy)

Col_Comparison

# Then the Output shows this. It is possible to turn off the Summary
# Messages, however:


Col_Comparison <- Column_Comparison(breakfast_df, breakfast_df_copy,
Summary_Message = FALSE)

Col_Comparison

}
