% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KE18_subset_modified}
\alias{KE18_subset_modified}
\title{KE19 Subset, modified to contain missing data}
\format{
A data frame with 21 rows and 9 columns:
\describe{
  \item{source_fct}{The FCT that this data is sourced from}
  \item{fdc_id}{food item ID}
  \item{food_desc}{Food Description}
  \item{food_group}{Food Group the food item belongs to}
  \item{ENERCkcal}{Energy in Kcal per 100g food}
  \item{WATERg}{Water in grams per 100g food}
  \item{PROCNTg}{Protein in grams per 100g food}
  \item{CHOAVLDFg}{Carbohydrates calculated by difference per 100g food}
  \item{VITB12mcg}{Vitamin B12 in micrograms per 100g of food}
}
}
\source{
A modified subset of the Kenya 2018 FCT, available here: https://openknowledge.fao.org/items/ece9c958-74f6-4d90-b14e-6b8a16c0dd9d
}
\usage{
KE18_subset_modified
}
\description{
A subset of the Kenya 2018 FCT dataset, with some data removed.
}
\keyword{datasets}
