% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutritionTools-package.R
\docType{package}
\name{NutritionTools-package}
\alias{NutritionTools}
\alias{NutritionTools-package}
\title{NutritionTools: Tools for the Organisation, Matching, Calculation, and Summarisation of Nutrition Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions useful for the organisation, matching, calculation, and summarising of Nutrition Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TomCodd/NutritionTools}
  \item \url{https://tomcodd.github.io/NutritionTools/}
  \item Report bugs at \url{https://github.com/TomCodd/NutritionTools/issues}
}

}
\author{
\strong{Maintainer}: Thomas Codd \email{tbc5@protonmail.com} (\href{https://orcid.org/0000-0002-4992-0591}{ORCID}) (https://github.com/TomCodd)

Authors:
\itemize{
  \item Lucia Segovia de la Revilla (\href{https://orcid.org/0000-0001-7712-2020}{ORCID}) (https://github.com/LuciaSegovia) [contributor]
}

}
\keyword{internal}
