% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SOP_example_df}
\alias{SOP_example_df}
\title{Imaginary Food Items}
\format{
A data frame with 10 rows and 9 columns:
\describe{
  \item{food_code}{Imaginary Food ID}
  \item{WATERg}{Imaginary Water values}
  \item{PROCNTg}{Imaginary Protein values}
  \item{FAT_g_standardised}{Imaginary Fat values}
  \item{CHOAVLDFg_standardised}{Imaginary Carbohyrate values}
  \item{FIBTGg_standardised}{Imaginary Fibre values}
  \item{ALCg}{Imaginary Alcohol values}
  \item{ASHg}{Imaginary Ash values}
  \item{comments}{Imaginary comments}
}
}
\source{
An imaginary dataset; no formal source.
}
\usage{
SOP_example_df
}
\description{
An imaginary dataset of Food Items used to demonstrate the SOPg_calculator function
}
\keyword{datasets}
