% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SOP_example_df_nonstandard}
\alias{SOP_example_df_nonstandard}
\title{Imaginary Food Items}
\format{
A data frame with 10 rows and 9 columns:
\describe{
  \item{food_code}{Imaginary Food ID}
  \item{Water_values_g}{Imaginary Water values}
  \item{PROCNT_values_g}{Imaginary Protein values}
  \item{FAT_values_g_standardised}{Imaginary Fat values}
  \item{CHOAVLDF_values_g_standardised}{Imaginary Carbohyrate values}
  \item{FIBTG_values_g_standardised}{Imaginary Fibre values}
  \item{ALC_values_g}{Imaginary Alcohol values}
  \item{ASH_values_g}{Imaginary Ash values}
  \item{comments_column}{Imaginary comments}
}
}
\source{
An imaginary dataset; no formal source.
}
\usage{
SOP_example_df_nonstandard
}
\description{
An imaginary dataset of Food Items used to demonstrate the SOPg_calculator function
}
\keyword{datasets}
