% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_Creator.R
\name{Table_Creator}
\alias{Table_Creator}
\title{Table Creator}
\usage{
Table_Creator(df, for_R_package = TRUE)
}
\arguments{
\item{df}{Required - The data.frame you are looking to recreate.}

\item{for_R_package}{Required - default: \code{TRUE} - Either \code{TRUE} or
\code{FALSE}. If set to \code{TRUE}, the output will include the
description framework for the data.frame, to be inserted into the data.R
file of the package.}
}
\value{
The code required to recreate a pre-existing table, and (optionally)
the description framework of the data.frame to be inserted into a packages
data.R file.
}
\description{
This function generates the code needed to create a pre-existing
data.frame from scratch.
}
