% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WA19_subset}
\alias{WA19_subset}
\title{WA19 subset}
\format{
A data frame with 49 rows and 9 columns:
\describe{
  \item{source_fct}{The FCT that this data is sourced from}
  \item{fdc_id}{food item ID}
  \item{food_desc}{Food Description}
  \item{food_group}{Food Group the food item belongs to}
  \item{ENERCkcal}{Energy in Kcal per 100g food}
  \item{WATERg}{Water in grams per 100g food}
  \item{PROCNTg}{Protein in grams per 100g food}
  \item{CHOAVLDFg}{Carbohydrates calculated by difference per 100g food}
  \item{VITB12mcg}{Vitamin B12 in micrograms per 100g of food}
}
}
\source{
A subset of the West Africa Food Composition  Table 2018 - found here: https://www.fao.org/infoods/infoods/tables-and-databases/faoinfoods-databases/en/
}
\usage{
WA19_subset
}
\description{
A subset of the West Africa 2019 FCT dataset.
}
\keyword{datasets}
