% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMatcher.R
\name{Fuzzy_Matcher}
\alias{Fuzzy_Matcher}
\title{A GUI interface to match rows in two dataframes to each other via a fuzzy
string search}
\usage{
Fuzzy_Matcher(df1, df2, focus_term)
}
\arguments{
\item{df1}{Required - The primary data frame, with items that need matches.
The first column must be the ID column, the second must be the item names.}

\item{df2}{Required - The secondary data frame, with a list of potential
items to match the contents of df1 against. The first column must be the ID
column, the second must be the item names.}

\item{focus_term}{Optional - Specify a string. If the string is contained
in the item name, then the fuzzy matcher opens a wider potential list of
matches to that item.}
}
\value{
An R object of csv that contains items from \code{df1} and their
counterparts from \code{df2} in the same row.
}
\description{
This function reads in two dataframes, both comprised of an ID
row and a name row. The name rows are matched based on fuzzy search
suggestions and human confirmation using the GUI interface.
}
